Fix value tests and make sure they run
