+++ src/lib.rs
@@ -61,6 +61,9 @@ pub use output_style::OutputStyle;
 
 pub use parser::{parse_scss_data, parse_scss_file};
 
+// FIXME: this should not be exposed, add a more sensible interface for this on parser
+pub use parser::value::value_expression;
+
 use selectors::Selectors;
 pub use unit::Unit;
 pub use value::{ListSeparator, Quotes, Value};
