+++ tests/value.rs
@@ -1,0 +1,7 @@
+extern crate num_rational;
+extern crate num_traits;
+extern crate nom;
+extern crate rsass;
+
+#[path="./value/mod.rs"]
+mod value;
