--- tests/value/parser_operations/mod.rs
@@ -1,5 +1,0 @@
-mod addition;
-mod binary_and_unary;
-mod logic_eq;
-
-use super::check_eval;
